requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
docarray==0.40.0
faiss-cpu==1.7.4
fastapi==0.110.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.8
langchain-cli==0.0.21
langchain-community==0.0.21
langchain-core==0.1.25
langchain-openai==0.0.6
langchainhub==0.1.14
langserve==0.0.45
langsmith==0.1.5
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.12.0
orjson==3.9.15
packaging==23.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
sse-starlette==1.8.2
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tomlkit==0.12.4
tqdm==4.66.2
typer==0.9.0
types-requests==2.31.0.20240218
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.23.2
yarl==1.9.4
